% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_rename}
\alias{col_rename}
\title{Adaptive column renaming}
\usage{
col_rename(df, old_name)
}
\arguments{
\item{df}{Input data.frame.}

\item{old_name}{Name of column to be renamed.}
}
\description{
This function renames columns in case the input data.frame includes any
colnames required by internal functions (e.g., \code{"y"}).
}
