% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_rename_all}
\alias{col_rename_all}
\title{Rename all problematic columns with col_rename().}
\usage{
col_rename_all(cn)
}
\arguments{
\item{Old}{column names.}
}
\value{
New columns names.
}
\description{
Rename all problematic columns with col_rename().
}
