% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_params.R
\name{correct_params}
\alias{correct_params}
\title{Function for correcting forde output with the function truncnorm_params.}
\usage{
correct_params(
  params,
  threshold = 1e-05,
  discard_deteriorations = T,
  keep_old_params = F,
  show_deviation = F,
  parallel = T
)
}
\arguments{
\item{params}{Circuit parameters learned via \code{\link{forde}}.}

\item{threshold}{Threshold value for deviations of mean and standard deviation
calculated from input parameters under which no correction will be performed.}

\item{discard_deteriorations}{Discard correction where sum of the deviation of
mean and standard deviation is higher than with input parameters?}

\item{keep_old_params}{Include input parameters in output?}

\item{parallel}{Compute in parallel? Must register backend beforehand, e.g.
via \code{doParallel}.}

\item{show_deviaton}{Include calculated deviations in output?}
}
\value{
Input circuit parameters with corrected mu and sigma.
}
\description{
Function for correcting forde output with the function truncnorm_params.
}
\examples{
# Train ARF and estimate leaf parameters
arf <- adversarial_rf(iris)
params <- forde(arf, iris)

# Correct params$cnt values for mu and sigma
params_corrected <- arf:::correct_params(params)

}
\keyword{internal}
