% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cparams2params}
\alias{cparams2params}
\title{Convert set of conditional delta params (\code{cparams}) to full \code{params}
set}
\usage{
cparams2params(params_base, cparams = NULL, parallel = TRUE)
}
\arguments{
\item{params_base}{Circuit parameters learned via \code{\link{forde}}.}

\item{cparams}{Set of conditional delta parameters based on \code{params_base}.}

\item{parallel}{Compute in parallel? Only relevant for \code{cparams} generated
with multi-row \code{evidence} and \code{row_mode = 'separate'}. Must
register backend beforehand, e.g. via \code{doParallel} or \code{doFuture};
see examples.}
}
\value{
Full circuit parameters conforming to \code{forde}-output.
For \code{cparams} generated with multi-row \code{evidence} and
\code{row_mode = 'separate'}, outputs a list of full circuit parameters.
}
\description{
This function converts a set of conditional params (output of cforde with
\code{output = 'cparams'}) to a full set of circuit parameters
}
\keyword{internal}
