% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Maximum A Posteriori Estimation}
\usage{
map(pc, query, evidence = NULL, n_eval = 100, parallel = FALSE)
}
\arguments{
\item{pc}{Probabilistic circuit learned via \code{\link{forde}}.}

\item{query}{Character vector of variable names. MAP estimates will be
computed for each.}

\item{evidence}{Optional set of conditioning events. This can take one of
three forms: (1) a partial sample, i.e. a single row of data with some but
not all columns; (2) a data frame of conditioning events, which allows for
inequalities; or (3) a posterior distribution over leaves. See Details.}

\item{n_eval}{Number of points to use for grid search.}

\item{parallel}{Compute in parallel? Must register backend beforehand, e.g.
via \code{doParallel}.}
}
\value{
A one row data frame with values for all query variables.
}
\description{
Compute the most likely value of some query variable(s), optionally
conditioned on some event(s).
}
\details{
This function computes maximum a posteriori (MAP) values for any subset of
features, optionally conditioned on some event(s). This technically covers
a range of related query types, including most probable explanations (MPE)
and marginal MAP, which are special instances of MAP.
}
\examples{
# Train ARF and corresponding circuit
arf <- adversarial_rf(iris)
psi <- forde(arf, iris)

# What is the most likely Sepal.Length?
map(psi, query = "Sepal.Length")

# What if we condition on Species = "setosa"?
evi <- data.frame(Species = "setosa")
map(psi, query = "Sepal.Length", evidence = evi)


}
\references{
Watson, D., Blesch, K., Kapar, J., & Wright, M. (2023). Adversarial random
forests for density estimation and generative modeling. In \emph{Proceedings
of the 26th International Conference on Artificial Intelligence and
Statistics}, pp. 5357-5375.
}
\seealso{
\code{\link{adversarial_rf}}, \code{\link{forde}}, \code{\link{lik}}
}
