% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_cond}
\alias{prep_cond}
\title{Preprocess conditions}
\usage{
prep_cond(condition, params, row_mode)
}
\arguments{
\item{condition}{Optional set of conditioning events.}

\item{params}{Circuit parameters learned via \code{\link{forde}}.}
}
\description{
This function prepares conditions for computing conditional circuit paramaters via cforde
}
