% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_evi}
\alias{prep_evi}
\title{Preprocess evidence}
\usage{
prep_evi(params, evidence)
}
\arguments{
\item{params}{Circuit parameters learned via \code{\link{forde}}.}

\item{evidence}{Optional set of conditioning events.}
}
\description{
This function prepares the evidence for computing leaf posteriors.
}
