% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_package.R
\name{read_RP4_samples_file}
\alias{read_RP4_samples_file}
\title{Read the input tsv file}
\usage{
read_RP4_samples_file(samplesFile)
}
\arguments{
\item{samplesFile}{"Samples.tsv" tab-seperated file that describes file locations and sample groupings.}
}
\description{
This function allows you to read the samples data, returning a dataframe.
}
\examples{
read_RP4_samples_file("samples.txt")
}
\keyword{Reader}
\keyword{Samples}
