% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample}
\alias{resample}
\title{Safer version of sample()}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{A vector of one or more elements from which to choose.}

\item{...}{Further arguments for sample().}
}
\value{
A vector of length size with elements drawn from x.
}
\description{
Safer version of sample()
}
