% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_package.R
\name{RP4_check_organism}
\alias{RP4_check_organism}
\title{Organism Check}
\usage{
RP4_check_organism(organism)
}
\arguments{
\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}
}
\description{
Check if the organism is valid.
}
