% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_package.R
\name{RP4_predict_heterogenity}
\alias{RP4_predict_heterogenity}
\title{Predict Differential Ribosomal Heterogeneity candidates with rRNA count data}
\usage{
RP4_predict_heterogenity(
  samples,
  rRNA_counts,
  compare = "group",
  organism = "hs",
  QCplot = FALSE,
  targetDir = NA,
  comparisons = NULL
)
}
\arguments{
\item{samples}{Samples dataframe created by read_samples_file() function.}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Direcory to save QC plots in.}

\item{comparisons}{List of comparisons to be included.}
}
\description{
Overlap differential rRNA count data with 3d ribosome, rRNA-RP proximity data and predict heterogeneity candidates across groups.
}
\examples{
RP4_predict_heterogenity(samples_df, rRNA_counts_df, organism="hs", QCplot=TRUE)
}
\keyword{Diffferential}
\keyword{Heterogeneity}
\keyword{RP}
\keyword{Ribosome}
\keyword{rRNA}
\keyword{ribosome}
