% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_package.R
\name{RP4_read_rRNA_fragments}
\alias{RP4_read_rRNA_fragments}
\title{Read rRNA quantification from bedGraph (maybe .bam as well) files}
\usage{
RP4_read_rRNA_fragments(
  samples,
  organism = "hs",
  QCplot = FALSE,
  targetDir = NA
)
}
\arguments{
\item{samples}{Samples dataframe created by read_samples_file() function.}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Direcory to save QC plots in.}
}
\description{
Read bedgraph files to create the rRNA count data.
}
\examples{
RP4_read_rRNA_fragments(samples_df, "hs")
}
\keyword{Reader}
\keyword{alignment}
\keyword{fragment}
\keyword{rRNA}
