% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_params.R
\name{truncnorm_params}
\alias{truncnorm_params}
\title{Function for calculating truncated normal parameters for given bounds and
desired mean and standard deviation.}
\usage{
truncnorm_params(a, b, mean, sd, parallel = TRUE)
}
\arguments{
\item{a}{Vector of lower bounds of truncated Gaussians.}

\item{b}{Vector of higher bounds of truncated Gaussians.}

\item{mean}{Vector of desired means of truncated Gaussians.}

\item{sd}{Vector of desired standard deviations of truncated Gaussians.}

\item{parallel}{Compute in parallel? Must register backend beforehand, e.g.
via \code{doParallel}.}
}
\value{
Parameter values for location and scale parameters for desired conditions.
}
\description{
Function for calculating truncated normal parameters for given bounds and
desired mean and standard deviation.
}
\keyword{internal}
