% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_params.R
\name{truncnorm_params_mle}
\alias{truncnorm_params_mle}
\title{Function for performing MLE for truncated Gaussians.}
\usage{
truncnorm_params_mle(x, a = -Inf, b = Inf)
}
\arguments{
\item{x}{Vector of input data.}

\item{a}{Vector of lower bounds of truncated Gaussians.}

\item{b}{Vector of higher bounds of truncated Gaussians.}
}
\value{
Parameter values for location and scale parameters for desired conditions.
}
\description{
Function for performing MLE for truncated Gaussians.
}
\keyword{internal}
