% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unoverlap_hyperrectangles}
\alias{unoverlap_hyperrectangles}
\title{Unoverlap hyperrectangles for or-ed conditions}
\usage{
unoverlap_hyperrectangles(hyperrectangles, cat_cols)
}
\arguments{
\item{hyperrectangles}{Input data.table of hyperrectangles in long format}

\item{cat_cols}{Vector of names of categorical features}
}
\description{
Algorithm is not optimal ( = number of resulting hyperrectangles is not minimal).
}
