% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_params.R
\name{var_truncnorm}
\alias{var_truncnorm}
\title{Quick-fix replacement function for vtruncnorm (because it seems to be buggy!)}
\usage{
var_truncnorm(a = -Inf, b = Inf, mean = 0, sd = 1)
}
\arguments{
\item{a}{Vector of lower bounds of truncated Gaussians.}

\item{b}{Vector of higher bounds of truncated Gaussians.}

\item{mean}{Vector of means of corresponding untruncated Gaussians.}

\item{sd}{Vector of standard deviations of corresponding untruncated Gaussians.}
}
\value{
Vector of variances of truncated Gaussians.
}
\description{
Quick-fix replacement function for vtruncnorm (because it seems to be buggy!)
}
\keyword{internal}
