% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.max.random}
\alias{which.max.random}
\title{which.max() with random at ties}
\usage{
which.max.random(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Index of maximum value in x, with random tie-breaking.
}
\description{
which.max() with random at ties
}
