% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF-package.R
\docType{package}
\name{ARF-package}
\alias{ARF}
\alias{ARF-package}
\title{ARF: ARF - Analysis of rRNA Fragments originating from Ribo-seq experiments}
\description{
ARF is a platform that consists multiple methods for the analysis of rRNA fragments coming from Ribo-seq experiments. It contains the following methods; dripARF, dricARF. dripARF predicts which Ribosomal Proteins contribute the differential ribosomal hetegeneity, whereas dricARF predicts abundance change in ribosome collisions across samples. Bot predictions are performed by overlapping rRNA fragment data with the 3D structure of different ribosome conformations.
}
\author{
\strong{Maintainer}: Ferhat Alkan \email{feralkan@gmail.com} (\href{https://orcid.org/0000-0001-6709-9605}{ORCID})

Authors:
\itemize{
  \item Edwin S Kwei-Baffour \email{e.kyei@nki.nl} (\href{https://orcid.org/0000-0002-2370-0451}{ORCID})
  \item William J Faller \email{w.faller@nki.nl} (\href{https://orcid.org/0000-0002-0738-2254}{ORCID})
}

}
