% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{ARF_check_organism}
\alias{ARF_check_organism}
\title{Organism Check}
\usage{
ARF_check_organism(organism)
}
\arguments{
\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}
}
\description{
Check if the organism is valid.
}
\examples{
ARF_check_organism("hs")
}
\keyword{ARF}
\keyword{Organism}
\keyword{dripARF}
