% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{ARF_convert_Ribo3D_pos}
\alias{ARF_convert_Ribo3D_pos}
\title{Convert 3D Ribosome distance file to target organism through rRNA alignments!}
\usage{
ARF_convert_Ribo3D_pos(
  source_distance_file,
  source_rRNAs_fasta,
  target_species,
  target_rRNAs_fasta,
  rRNA_pairs = list(),
  source_positions = NULL,
  type = "distances"
)
}
\arguments{
\item{source_distance_file}{XXX.ARF.minimum_distances.csv file generated with ARF_parse_PDB_ribosome() for the source 3D Ribosome}

\item{source_rRNAs_fasta}{Fasta file for the rRNAs of the source organism.}

\item{target_species}{ID for the target species, i.e. mm, sc, etc.}

\item{target_rRNAs_fasta}{Fasta file for the rRNAs of the target organism. Same file used in rRNA fragment alignment.}

\item{rRNA_pairs}{List of rRNA ID pairs matching source and target rRNAs i.e. list(c("human_28S","mouse_28S"), etc. ).}

\item{source_positions}{List of position vectors to convert (Alternative input)}

\item{type}{Which conversion to perform (Currently distances or positionset only)}
}
\description{
Align rRNAs from source and target organism, then, convert source species 3D ribosome data accordingly
}
\examples{
ARF_convert_ribosome3D_rRNA_pos()
}
\keyword{3D}
\keyword{PDB}
\keyword{analysis}
\keyword{file}
\keyword{ribosome}
\keyword{using}
