% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{ARF_parse_PDB_ribosome}
\alias{ARF_parse_PDB_ribosome}
\title{Parse 3D Ribosome data from PDB!}
\usage{
ARF_parse_PDB_ribosome(
  species,
  PDBid,
  PDB_file = NULL,
  rRNAs_file = NULL,
  PDB_chains_2_RP_nomenclature = NULL,
  download_directory = getwd(),
  method = "dripARF_minDistances",
  out_prefix = "Ribosome.3D",
  duplication_resolve = min
)
}
\arguments{
\item{species}{ID for the species to be analyzed, i.e. human}

\item{PDBid}{ID for the 3D structure, i.e. "4V6X" for human}

\item{PDB_file}{CIF file for the 3D ribosome, optional}

\item{rRNAs_file}{rRNAs file for the 3D ribosome, (must have been generated by ARF) optional}

\item{PDB_chains_2_RP_nomenclature}{Alternative data.frame that maps chain IDs to RP names and new nomenclature. Required columns are ID, RP_name, RP_new, chainID.}

\item{download_directory}{Directory for downloads, default=getwd()}

\item{method}{Which analysis to perform , default:"dripARF_minDistances" which calculates rRNA-RP min distances. (No other option yet)}

\item{duplication_resolve}{How do we resolve duplicated chains (Default take the min distance)}

\item{outprefix}{Prefix for the file to save results within, default: "Ribosome.3D"}
}
\description{
Parse 3D ribosome data
}
\examples{
ARF_parse_PDB_ribosome("4V6X")
}
\keyword{3D}
\keyword{PDB}
\keyword{analysis}
\keyword{file}
\keyword{ribosome}
\keyword{using}
