% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cforde}
\alias{cforde}
\title{Compute conditional circuit parameters}
\usage{
cforde(
  params,
  condition,
  row_mode = c("separate", "or"),
  stepsize = 200,
  parallel = T
)
}
\arguments{
\item{params}{Circuit parameters learned via \code{\link{forde}}.}

\item{condition}{Data frame of conditioning event(s).}
}
\description{
Compute conditional circuit parameters
}
