% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_rename}
\alias{col_rename}
\title{Adaptive column renaming}
\usage{
col_rename(cn, old_name)
}
\arguments{
\item{cn}{Column names.}

\item{old_name}{Name of column to be renamed.}
}
\description{
This function renames columns in case the input colnames includes any
colnames required by internal functions (e.g., \code{"y"}).
}
\keyword{internal}
