% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dricARF.R
\name{dricARF}
\alias{dricARF}
\title{dricARF wrapper}
\usage{
dricARF(
  samplesFile,
  rRNAs_fasta,
  samples_df = NULL,
  organism = NULL,
  compare = "group",
  QCplot = TRUE,
  targetDir = NA,
  comparisons = NULL,
  exclude = NULL,
  GSEAplots = FALSE,
  gsea_sets_RP = NULL,
  RP_proximity_df = NULL,
  gsea_sets_Collision = NULL
)
}
\arguments{
\item{samplesFile}{File that describes file locations and sample groupings}

\item{rRNAs_fasta}{Fasta file for the 4 rRNAs of the organism.}

\item{organism}{Organism abbrevation. Pass "hs" for human, "mm" for mouse, and "sc" for yeast.}

\item{compare}{If you want to compare samples based on other grouping, choose the columnname that is given in samplesFile (Default=group).}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in.}

\item{comparisons}{List of comparisons to be included.}

\item{exclude}{List of sample names to be excluded from the analysis.}

\item{GSEAplots}{Whether to produce and save the standard GSEA plots.}

\item{gsea_sets_RP}{RP-rRNA contact point sets to perform dripARF enrichments on. (preset for hs, mm, and sc predictions)}

\item{RP_proximity_df}{RP-rRNA proximity matrix that is calculated by ARF. (preset for hs, mm, and sc predictions)}

\item{gsea_sets_Collision}{Ribosome collision sets to perform dricARF enrichments on. (preset for hs, mm, and sc predictions)}

\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}
}
\description{
This function allows you to run the whole dricARF pipeline
}
\examples{
dricARF("samples.txt", "rRNAs.fa", organism="mm", targetDir="/target/directory/to/save/results")
}
\keyword{dricARF}
\keyword{pipeline}
