% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dricARF.R
\name{dricARF_liftover_collision_sets}
\alias{dricARF_liftover_collision_sets}
\title{Convert 3D Ribosome distance file to target organism through rRNA alignments!}
\usage{
dricARF_liftover_collision_sets(
  target_species,
  target_rRNAs_fasta,
  rRNA_pairs = list()
)
}
\arguments{
\item{target_species}{ID for the target species, i.e. mm, sc, etc.}

\item{target_rRNAs_fasta}{Fasta file for the rRNAs of the target organism. Same file used in rRNA fragment alignment.}

\item{rRNA_pairs}{List of rRNA ID pairs matching source and target rRNAs. (Use 28S, 18S, 5.8S, 5S as source ids) i.e. list(c("28S","species_28S"), c("18S","species_18S") etc.).}
}
\description{
Align rRNAs from hs/sc and target organism, then, convert the collision sets into target organism coordinates.
}
\examples{
dricARF_liftover_collision_sets("dm","drosophila_rRNAs.fa",rRNA_pairs=list(c("28S","dm_rRNA_28S"),c("18S","dm_rRNA_18S"),c("5.8S","dm_rRNA_5.8S"),c("5S","dm_rRNA_5S")))
}
\keyword{3D}
\keyword{PDB}
\keyword{analysis}
\keyword{file}
\keyword{ribosome}
\keyword{using}
