% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dricARF.R
\name{dricARF_result_scatterplot}
\alias{dricARF_result_scatterplot}
\title{Draw dricARF result scatterplot for multiple comparisons}
\usage{
dricARF_result_scatterplot(
  dricARF_results,
  targetDir,
  title = "dricARF (highlighted) & dripARF predictions",
  addedRPs = NULL,
  highlightRPs = NULL,
  randZscore_thr = 1,
  ORA_adjP_thr = 0.1,
  RPSEA_adjP_thr = 0.1,
  ORA_sig_n = 1
)
}
\arguments{
\item{dricARF_results}{Full dripARF result data.}

\item{targetDir}{Directory to save the plots in.}

\item{title}{Default is "dricARF (highlighted) & dripARF predictions"}

\item{addedRPs}{Add given RPs to the plot no matter if they are significant or not.}

\item{highlightRPs}{List of RPs to highlight instead of highlighting the top-predicted RPs.}

\item{randZscore_thr}{Default=1}

\item{ORA_adjP_thr}{Default=0.05}

\item{RPSEA_adjP_thr}{Default=0.05}

\item{ORA_sig_n}{Default=1}
}
\description{
Draw scatterplot for GSEA and ORA analyses
}
\examples{
dripARF_result_scatterplot(dripARF_results, "/Folder/to/save/in/")
}
\keyword{Collisions}
\keyword{Differential}
\keyword{Ribosome}
\keyword{dricARF}
\keyword{scatterplot}
