% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF}
\alias{dripARF}
\title{dripARF wrapper}
\usage{
dripARF(
  samplesFile,
  organism = "hs",
  compare = "group",
  QCplot = TRUE,
  targetDir = NA,
  comparisons = NULL,
  exclude = NULL,
  GSEAplots = FALSE
)
}
\arguments{
\item{samplesFile}{File that describes file locations and sample groupings}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{compare}{If you want to compare samples based on other grouping, choose the columnname that is given in samplesFile (Default=group).}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in.}

\item{comparisons}{List of comparisons to be included.}

\item{exclude}{List of sample names to be excluded from the analysis.}

\item{GSEAplots}{Whether to produce and save the standard GSEA plots.}
}
\description{
This function allows you to run the whole dripARF pipeline
}
\examples{
dripARF("samples.txt", organism="mm", targetDir="/target/directory/to/save/results")
}
\keyword{dripARF}
\keyword{pipeline}
