% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_add_replicates}
\alias{dripARF_add_replicates}
\title{Add random-edited replicates to single replicate groups}
\usage{
dripARF_add_replicates(samples, rRNA_counts, QCplot = FALSE, targetDir = NA)
}
\arguments{
\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}

\item{rRNA_counts}{rRNA_counts that were read by dripARF_read_rRNA_fragments() function}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in. (Default: working directory, getwd() output)}
}
\description{
Add random-edited replicates to single replicate groups
}
\examples{
dripARF_add_replicates(samples_df)
dripARF_add_replicates(samples_df, QCplot=TRUE, targetDir="./")
}
\keyword{Reader}
\keyword{alignment}
\keyword{fragment}
\keyword{rRNA}
