% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_get_DESEQ_dds}
\alias{dripARF_get_DESEQ_dds}
\title{To normalize the counts with DESEQ}
\usage{
dripARF_get_DESEQ_dds(
  samples,
  rRNA_counts = NULL,
  compare = "group",
  organism = "hs",
  exclude = NULL,
  count_threshold = 100,
  QCplot = FALSE,
  targetDir = NA
)
}
\arguments{
\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}

\item{rRNA_counts}{rRNA_counts that were read by dripARF_read_rRNA_fragments() function (optional)}

\item{compare}{If you want to compare samples based on other grouping, choose the columnname that is given in the samplesFile (Default=group).}

\item{organism}{Organism abbrevation. Pass "hs" for human, "mm" for mouse, and "sc" for yeast.}

\item{exclude}{List of sample names to be excluded from the analysis.}

\item{count_threshold}{Exclude the positions with reads less than this threshold on average. (Default: 1000)}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in. (Default: working directory, getwd() output)}
}
\description{
A function that normalizes rRNA counts with DESEQ2
}
\examples{
dripARF_get_DESEQ_dds(samples_df)
}
\keyword{Normalization}
