% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dripARF.R
\name{dripARF_get_RP_proximity_sets}
\alias{dripARF_get_RP_proximity_sets}
\title{dripARF threshold_test wrapper}
\usage{
dripARF_get_RP_proximity_sets(
  RP_proximity_df,
  additional_RPcols = c(),
  rRNAs_fasta = NULL,
  thresholds = NULL
)
}
\arguments{
\item{RP_proximity_df}{RP-rRNA proximity matrix that is calculated by ARF.}

\item{additional_RPcols}{If you have added your own columns to RP_proximity_df, please provide their column indexes.}

\item{rRNAs_fasta}{Fasta file for the 4 rRNAs of the target organism (target rRNAs file for converted positions).}

\item{thresholds}{Thresholds for creating the RP-specific rRNA proximity sets.}
}
\description{
This function allows you to run the whole dripARF pipeline with varying proximity thresholds
}
\examples{
dripARF_get_RP_proximity_sets(RP_proximities_dataframe, rRNAs_fasta="rRNAs.fa")
dripARF_get_RP_proximity_sets(RP_proximities_dataframe, additional_RPcols=80:95, rRNAs_fasta="rRNAs.fa")
}
\keyword{RP}
\keyword{RPSEA}
\keyword{RRNA}
\keyword{proximity}
\keyword{set}
