% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_predict_heterogenity}
\alias{dripARF_predict_heterogenity}
\title{Predict Differential Ribosomal Heterogeneity candidates with rRNA count data}
\usage{
dripARF_predict_heterogenity(
  samples,
  rRNA_counts = NULL,
  dripARF_dds = NULL,
  compare = "group",
  organism = "hs",
  QCplot = FALSE,
  targetDir = NA,
  comparisons = NULL,
  exclude = NULL,
  GSEAplots = FALSE
)
}
\arguments{
\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}

\item{rRNA_counts}{rRNA_counts that were read by dripARF_read_rRNA_fragments() function: (optional)}

\item{dripARF_dds}{DESEQ2 normalized rRNA_counts coming from dripARF_get_DESEQ_dds() function: (optional)}

\item{compare}{If you want to compare samples based on other grouping, choose the columnname that is given in samplesFile (Default=group).}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QCplots in.}

\item{comparisons}{List of comparisons to be included.}

\item{exclude}{List of sample names to be excluded from the analysis.}

\item{GSEAplots}{Whether to produce standard GSEA plots.}
}
\description{
Overlap differential rRNA count data with 3d ribosome, rRNA-RP proximity data and predict heterogeneity candidates across groups.
}
\examples{
dripARF_predict_heterogenity(samples_df, rRNA_counts_df, organism="hs", QCplot=TRUE)
}
\keyword{Diffferential}
\keyword{Heterogeneity}
\keyword{RP}
\keyword{Ribosome}
\keyword{rRNA}
\keyword{ribosome}
