% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_read_rRNA_fragments}
\alias{dripARF_read_rRNA_fragments}
\title{Read rRNA quantification from .bedGraph or .bam files}
\usage{
dripARF_read_rRNA_fragments(
  samples,
  organism = "hs",
  QCplot = FALSE,
  targetDir = NA
)
}
\arguments{
\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in. (Default: working directory, getwd() output)}
}
\description{
Read bedgraph/bam files to create the rRNA count data.
}
\examples{
dripARF_read_rRNA_fragments(samples_df, organism="hs")
dripARF_read_rRNA_fragments(samples_df, organism="hs", QCplot=TRUE, targetDir="./")
}
\keyword{Reader}
\keyword{alignment}
\keyword{fragment}
\keyword{rRNA}
