% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dripARF.R
\name{dripARF_read_rRNA_fragments}
\alias{dripARF_read_rRNA_fragments}
\title{Read rRNA quantification from .bedGraph or .bam files}
\usage{
dripARF_read_rRNA_fragments(
  samples,
  rRNAs_fasta,
  organism = NULL,
  QCplot = FALSE,
  targetDir = NA
)
}
\arguments{
\item{samples}{Samples dataframe created by read_ARF_samples_file() function.}

\item{rRNAs_fasta}{Fasta file for the 4 rRNAs of the organism.}

\item{organism}{Organism abbrevation. Pass "hs" for human, "mm" for mouse, and "sc" for yeast.}

\item{QCplot}{TRUE or FALSE, whether to generate QC plots or not.}

\item{targetDir}{Directory to save the QC plots in. (Default: working directory, getwd() output)}

\item{add_random_replicates}{For sample gorups with a single replicate, add a second one that is "randomly" altered (Default: F)}
}
\description{
Read bedgraph/bam files to create the rRNA count data.
}
\examples{
dripARF_read_rRNA_fragments(samples_df, "rRNAs.fa", organism="hs")
dripARF_read_rRNA_fragments(samples_df, "rRNAs.fa", organism="hs", QCplot=TRUE, targetDir="./")
}
\keyword{Reader}
\keyword{alignment}
\keyword{fragment}
\keyword{rRNA}
