% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_result_heatmap}
\alias{dripARF_result_heatmap}
\title{Draw heatmap for multiple comparisons.}
\usage{
dripARF_result_heatmap(
  dripARF_results,
  title,
  targetDir,
  addedRPs = NULL,
  randZscore_thr = c(1, 1, 1.5, 0),
  ORA_adjP_thr = c(0.05, 2, 0.05, 2),
  RPSEA_adjP_thr = c(0.05, 0.05, 0.05, 0.05),
  ORA_sig_n = c(1, 1, 1, 1)
)
}
\arguments{
\item{dripARF_results}{dripARF result data.}

\item{title}{Title for the plot and files.}

\item{targetDir}{Directory to save the plots in.}

\item{addedRPs}{Add given RPs to the plot no matter if they are significant or not..}
}
\description{
Draw heatmap of NES and NES_randZscore based on different thresholds.
}
\examples{
dripARF_result_heatmap(dripARF_results,"Title","/Folder/to/save/in/")
dripARF_result_heatmap(dripARF_results,"Title","/Folder/to/save/in/", randZscore_thr=1)
}
\keyword{Differential}
\keyword{Heatmap}
\keyword{Heterogeneity}
\keyword{Ribosome}
