% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_result_scatterplot}
\alias{dripARF_result_scatterplot}
\title{Draw dripARF result scatterplot for multiple comparisons}
\usage{
dripARF_result_scatterplot(
  dripARF_results,
  targetDir,
  title = "DRH_prediction_volcanos",
  addedRPs = NULL,
  highlightRPS = NULL,
  randZscore_thr = 1,
  ORA_adjP_thr = 0.05,
  RPSEA_adjP_thr = 0.05,
  ORA_sig_n = 1
)
}
\arguments{
\item{dripARF_results}{Full dripARF result data.}

\item{targetDir}{Directory to save the plots in.}

\item{title}{Default is "DRH_prediction_volcanos"#'}

\item{addedRPs}{Add given RPs to the plot no matter if they are significant or not.}

\item{highlightRPS}{List of RPs to highlight instead of highlighting the top RPs.}

\item{randZscore_thr}{Default=1}

\item{ORA_adjP_thr}{Default=0.05}

\item{RPSEA_adjP_thr}{Default=0.05}

\item{ORA_sig_n}{Default=1}
}
\description{
Draw scatterplot for GSEA and ORA analyses
}
\examples{
dripARF_result_scatterplot(dripARF_results,"/Folder/to/save/in/")
}
\keyword{Differential}
\keyword{Heatmap}
\keyword{Heterogeneity}
\keyword{Ribosome}
