% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_simplify_results}
\alias{dripARF_simplify_results}
\title{Simplify result file}
\usage{
dripARF_simplify_results(
  dripARF_results,
  randZscore_thr = 1,
  ORA_adjP_thr = 0.05,
  RPSEA_adjP_thr = 0.05,
  ORA_sig_n = 0
)
}
\arguments{
\item{dripARF_results}{dripARF results data frame}

\item{randZscore_thr}{Default=1}

\item{ORA_adjP_thr}{Default=0.05}

\item{RPSEA_adjP_thr}{Default=0.05}
}
\description{
Simplify results based on thresholds
}
\examples{
dripARF_simplify_results(dripARF_results)
}
\keyword{dripARF}
\keyword{results}
\keyword{simple}
