% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{dripARF_threshold_test}
\alias{dripARF_threshold_test}
\title{dripARF threshold_test wrapper}
\usage{
dripARF_threshold_test(
  samplesFile,
  organism = "hs",
  compare = "group",
  comparisons = NULL,
  exclude = NULL,
  thresholds = NULL
)
}
\arguments{
\item{samplesFile}{File that describes file locations and sample groupings}

\item{organism}{Organism abbrevation. Pass "hs" for human and "mm" for mouse.}

\item{compare}{If you want to compare samples based on other grouping, choose the columnname that is given in samplesFile (Default=group).}

\item{comparisons}{List of comparisons to be included.}

\item{exclude}{List of sample names to be excluded from the analysis.}

\item{thresholds}{List of given threshold vectors. Every vector contains the Angstrom threshold as 1st value, Set-size threshold as 2nd, and directionality as 3rd (optional, default=FALSE).}
}
\description{
This function allows you to run the whole dripARF pipeline with varying proximity thresholds
}
\examples{
dripARF_threshold_test("samples.txt", organism="mm", thresholds=list(c(20,150),c(15,100)))

}
\keyword{dripARF}
\keyword{pipeline}
