% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{leaf_posterior}
\alias{leaf_posterior}
\title{Compute leaf posterior}
\usage{
leaf_posterior(params, evidence)
}
\arguments{
\item{params}{Circuit parameters learned via \code{\link{forde}}.}

\item{evidence}{Data frame of conditioning event(s).}
}
\description{
This function returns a posterior distribution on leaves, conditional on some
evidence.
}
