% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_x}
\alias{post_x}
\title{Post-process data}
\usage{
post_x(x, params)
}
\arguments{
\item{x}{Input data.frame.}

\item{params}{Circuit parameters learned via \code{\link{forde}}.}
}
\description{
This function prepares output data for map and forge.
}
