% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_cond}
\alias{prep_cond}
\title{Preprocess conditions}
\usage{
prep_cond(evidence, params, row_mode)
}
\arguments{
\item{evidence}{Optional set of conditioning events.}

\item{params}{Circuit parameters learned via \code{\link{forde}}.}

\item{row_mode}{Interpretation of rows in multi-row conditions.}
}
\description{
Preprocess conditions
}
\keyword{internal}
