% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARF_platform.R
\name{read_ARF_samples_file}
\alias{read_ARF_samples_file}
\title{Read the input tsv file}
\usage{
read_ARF_samples_file(samplesFile)
}
\arguments{
\item{samplesFile}{File that describes file locations and sample groupings.}
}
\description{
This function allows you to read the samples data, returning a dataframe.
}
\examples{
read_ARF_samples_file("samples.txt")
}
\keyword{Reader}
\keyword{Samples}
