% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RP4_package.R
\name{RP4}
\alias{RP4}
\title{RP4 wrapper}
\usage{
RP4(samplesFile, organism = "hs", QCplot = TRUE, targetDir = NA)
}
\arguments{
\item{samplesFile}{Samples.txt file that describes file locations and sample groupings}
}
\description{
This function allows you to run the whole RP4 pipeline
}
\examples{
RP4("samples.txt")
}
\keyword{RP4}
\keyword{pipeline}
