% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_from_leaves.R
\name{sample_from_leaves}
\alias{sample_from_leaves}
\title{Generate synthetic data by sampling from leaves of a random forest}
\usage{
sample_from_leaves(arf, x_real, factor_cols = NULL, lvls = NULL, prep = TRUE)
}
\arguments{
\item{arf}{Random forest to sample from}

\item{x_real}{Data to sample from}

\item{factor_cols}{Logical vector indicating which columns are factors (optional)}

\item{lvls}{List of factor levels (optional)}
}
\value{
A data.table of synthetic data
}
\description{
Generate synthetic data by sampling from leaves of a random forest
}
\examples{
arf <- adversarial_rf(iris)
sample_from_leaves(arf, iris)
}
